+++ scripts/add/1-map-names-to-teams.ts
@@ -1,8 +1,7 @@
 import * as fs from "fs";
-import * as path from "path";
-import { FILE_PATH, GENDER, ObjectId } from "./0-settings";
+import { FILE_PATH, GENDER, ObjectId } from "./settings";
 
-const list = String(fs.readdirSync(FILE_PATH)).split("\n");
+const list = String(fs.readFileSync(FILE_PATH)).split("\n");
 
 const teams = list.map(team => {
   return {
