+++ scripts/add/3-get-teams-for-division.ts
@@ -1,6 +1,4 @@
-import * as fs from "fs";
-import * as path from "path";
-import { FILE_PATH } from "./0-settings";
+import { FILE_PATH } from "./settings";
 
 const teams = require(FILE_PATH);
