+++ scripts/add/set-id.ts
@@ -1,0 +1,8 @@
+import { ObjectId } from "./settings";
+
+console.log(ObjectId());
+console.log(ObjectId());
+console.log(ObjectId());
+console.log(ObjectId());
+
+console.log(ObjectId());
