+++ scripts/data/clubs.json
@@ -511,5 +511,41 @@
     "owner": {
       "$oid": "598f07679a5d6fdee4ae8bcd"
     }
+  },
+  {
+    "_id": {
+      "$oid": "598f7d3e35b61604a6fbf5a6"
+    },
+    "name": "JRX",
+    "owner": {
+      "$oid": "598f07679a5d6fdee4ae8bcd"
+    }
+  },
+  {
+    "_id": {
+      "$oid": "598f7d3eccb6894bac4babd7"
+    },
+    "name": "Red Leicester",
+    "owner": {
+      "$oid": "598f07679a5d6fdee4ae8bcd"
+    }
+  },
+  {
+    "_id": {
+      "$oid": "598f7d3e6742de9ba15274cc"
+    },
+    "name": "GB U20",
+    "owner": {
+      "$oid": "598f07679a5d6fdee4ae8bcd"
+    }
+  },
+  {
+    "_id": {
+      "$oid": "598f7d3e9535467162e83fd9"
+    },
+    "name": "Irish Junior",
+    "owner": {
+      "$oid": "598f07679a5d6fdee4ae8bcd"
+    }
   }
 ]
