+++ scripts/data/tour1/divisions.json
@@ -1,0 +1,299 @@
+[
+  {
+    "gender": "Mens",
+    "name": "A Tour",
+    "tournament": {
+      "$oid": "598f3112e85028a6a6c7f62b"
+    },
+    "_id": {
+      "$oid": "598f5b382846d9574146d0f9"
+    },
+    "teams": [
+      {
+        "$oid": "598f628b3e70c85b7aabab95"
+      },
+      {
+        "$oid": "598f628bd9c6f9ed00717782"
+      },
+      {
+        "$oid": "598f628b6cc32bf62dedc035"
+      },
+      {
+        "$oid": "598f628bba3ddf8f9692c9d4"
+      },
+      {
+        "$oid": "598f628b6504df3e310dfa4f"
+      },
+      {
+        "$oid": "598f628b346417e2466d64de"
+      },
+      {
+        "$oid": "598f628b2c1769c31cd95ece"
+      },
+      {
+        "$oid": "598f628b9d51f10945fb6638"
+      },
+      {
+        "$oid": "598f628bc713a93b6a2ccff7"
+      },
+      {
+        "$oid": "598f628b5c7f31db6debd68a"
+      },
+      {
+        "$oid": "598f628b23b5b06037ecd524"
+      },
+      {
+        "$oid": "598f628b60148f3a8424bff9"
+      },
+      {
+        "$oid": "598f628bd12b845cc2559cdd"
+      },
+      {
+        "$oid": "598f628b44381ccd0c41aa83"
+      },
+      {
+        "$oid": "598f628bb2e520b5ff272e42"
+      },
+      {
+        "$oid": "598f628b009bd3c7841f1432"
+      }
+    ]
+  },
+  {
+    "gender": "Mens",
+    "name": "B Tour",
+    "tournament": {
+      "$oid": "598f3112e85028a6a6c7f62b"
+    },
+    "_id": {
+      "$oid": "598f5b38d880ed9e9cbb0245"
+    },
+    "teams": [
+      {
+        "$oid": "598f639964e2e141fb8c50fc"
+      },
+      {
+        "$oid": "598f63999ca2ddeba8a87a7e"
+      },
+      {
+        "$oid": "598f63994fa75e8f057262dd"
+      },
+      {
+        "$oid": "598f6399a74153fd57182902"
+      },
+      {
+        "$oid": "598f6399a0b6113558c3724b"
+      },
+      {
+        "$oid": "598f6399a2b66dbd143bc992"
+      },
+      {
+        "$oid": "598f63999371389b36350cf6"
+      },
+      {
+        "$oid": "598f639999180d8962b077d5"
+      },
+      {
+        "$oid": "598f63997a20e47043c2cf83"
+      },
+      {
+        "$oid": "598f639966aa01181bf6dfee"
+      },
+      {
+        "$oid": "598f6399df4d32cfed9374d8"
+      },
+      {
+        "$oid": "598f6399fa9c398fec808b68"
+      },
+      {
+        "$oid": "598f63992243e15d2e3b2213"
+      },
+      {
+        "$oid": "598f63998d691ec2dec8a89c"
+      },
+      {
+        "$oid": "598f6399d23f179ba45454f5"
+      },
+      {
+        "$oid": "598f6399d12b244dac9036e6"
+      }
+    ]
+  },
+  {
+    "gender": "Mens",
+    "name": "C Tour",
+    "tournament": {
+      "$oid": "598f3112e85028a6a6c7f62b"
+    },
+    "_id": {
+      "$oid": "598f5b3831207e89b2e6f406"
+    },
+    "teams": [
+      {
+        "$oid": "598f67e0f72a23bd41011129"
+      },
+      {
+        "$oid": "598f67e0f60e823e3a03bde5"
+      },
+      {
+        "$oid": "598f67e0b8c95d6f288c3388"
+      },
+      {
+        "$oid": "598f67e07bae2677a986e8d4"
+      },
+      {
+        "$oid": "598f67e0cc9a79684ca2cb68"
+      },
+      {
+        "$oid": "598f67e065f5c3f1b727e42c"
+      },
+      {
+        "$oid": "598f67e00a73acf86e19749e"
+      },
+      {
+        "$oid": "598f67e08d4dcc6c087ddfb2"
+      },
+      {
+        "$oid": "598f67e06793c616c29f4021"
+      },
+      {
+        "$oid": "598f67e07b77d0126edb4c23"
+      },
+      {
+        "$oid": "598f67e0f8253ee8327ed724"
+      },
+      {
+        "$oid": "598f67e0df0461ce4f8624ae"
+      },
+      {
+        "$oid": "598f67e04143dc0fb0645cdb"
+      },
+      {
+        "$oid": "598f67e034b660a781c2f314"
+      },
+      {
+        "$oid": "598f67e0f9c9026b731bcca2"
+      },
+      {
+        "$oid": "598f67e02ba87b03d4f34db5"
+      },
+      {
+        "$oid": "598f67e08c0cd462837fad79"
+      },
+      {
+        "$oid": "598f67e02157a98b8f7e4e58"
+      },
+      {
+        "$oid": "598f67e0da83a32c67c640e8"
+      },
+      {
+        "$oid": "598f67e0faf9ecabb253fc42"
+      },
+      {
+        "$oid": "598f67e09ffc1062ad47d67e"
+      },
+      {
+        "$oid": "598f67e0561075fd74775b65"
+      },
+      {
+        "$oid": "598f67e0035abc355562086b"
+      },
+      {
+        "$oid": "598f67e0fbacc92c8310103e"
+      },
+      {
+        "$oid": "598f67e07dc0fbd74b3fd8f9"
+      },
+      {
+        "$oid": "598f67e0e0dded8d2015e130"
+      },
+      {
+        "$oid": "598f67e03f4f2f645a748306"
+      },
+      {
+        "$oid": "598f67e0124dc438ae1cc8d4"
+      }
+    ]
+  },
+  {
+    "gender": "Womens",
+    "name": "Womens",
+    "tournament": {
+      "$oid": "598f3112e85028a6a6c7f62b"
+    },
+    "_id": {
+      "$oid": "598f5b38ff8e1b6368342a62"
+    },
+    "teams": [
+      {
+        "$oid": "598f7136d8de5fee8af532dd"
+      },
+      {
+        "$oid": "598f71360f750469ee6820e5"
+      },
+      {
+        "$oid": "598f713625dfa6abba7dbb9f"
+      },
+      {
+        "$oid": "598f71362d6fbef88fe93234"
+      },
+      {
+        "$oid": "598f7136d9151b44417503d5"
+      },
+      {
+        "$oid": "598f7136f04b3f37dcc4d768"
+      },
+      {
+        "$oid": "598f71367b6148338b3210c6"
+      },
+      {
+        "$oid": "598f713607173a093385c390"
+      },
+      {
+        "$oid": "598f71368248227249fc2940"
+      },
+      {
+        "$oid": "598f7136519dd064f398bfec"
+      },
+      {
+        "$oid": "598f713698b6dc7b2a8a45da"
+      },
+      {
+        "$oid": "598f7136a17cba49fa5bbd5b"
+      },
+      {
+        "$oid": "598f7136508f6884ddc85502"
+      },
+      {
+        "$oid": "598f71369dae909cddf7378f"
+      },
+      {
+        "$oid": "598f713617bf3d4216e1de60"
+      },
+      {
+        "$oid": "598f713643300b1a37f98985"
+      },
+      {
+        "$oid": "598f7136e6b87b42ca3e0ccb"
+      },
+      {
+        "$oid": "598f7136346d499fdbd7d44a"
+      },
+      {
+        "$oid": "598f71361193dbc74798ec2a"
+      },
+      {
+        "$oid": "598f71363de5bccb4cd13473"
+      },
+      {
+        "$oid": "598f71363a8f1806cc06e8c6"
+      },
+      {
+        "$oid": "598f713612a2c52826df866f"
+      },
+      {
+        "$oid": "598f71364ccad83f2b33afae"
+      }
+    ]
+  }
+]
