+++ scripts/data/tour1/tournament.json
@@ -1,0 +1,33 @@
+[
+  {
+    "name": "UKU Nottingham WindFarm 2017",
+    "location": "Nottingham",
+    "description": "UKU Nottingham WindFarm, Open & Women's Tour 1 2017 will be held at Grove Farm, University of Nottingham, with all divisions of the UKU Tour in one location.",
+    "_id": {
+      "$oid": "598f3112e85028a6a6c7f62b"
+    },
+    "owner": {
+      "$oid": "598f07679a5d6fdee4ae8bcd"
+    },
+    "startDate": {
+      "$date": "2017-07-02T23:00:00.000Z"
+    },
+    "endDate": {
+      "$date": "2017-07-04T22:59:59.999Z"
+    },
+    "divisions": [
+      {
+        "$oid": "598f5b382846d9574146d0f9"
+      },
+      {
+        "$oid": "598f5b38d880ed9e9cbb0245"
+      },
+      {
+        "$oid": "598f5b3831207e89b2e6f406"
+      },
+      {
+        "$oid": "598f5b38ff8e1b6368342a62"
+      }
+    ]
+  }
+]
