+++ scripts/data/tour2/teams-a-s-men.json
@@ -1,16 +1,162 @@
-Clapham A
-EMO 1
-Reading 1
-Clapham 1
-Glasgow
-Chevron Action Flash
-Devon 1
-Ireland Under-24 Men
-Fire of London 1
-Manchester 1
-GB U24 Darwin
-GB U24 Brisbane
-SMOG
-Flump
-Ka-Pow!
-Brighton City
+[
+  {
+    "_id": {
+      "$oid": "598f7743b3d5ac61de0d2714"
+    },
+    "club": {
+      "$oid": "598f59b96081bab2d6d7409b"
+    },
+    "gender": "Mens",
+    "name": "Clapham A"
+  },
+  {
+    "_id": {
+      "$oid": "598f77431537ac48d5694c2e"
+    },
+    "club": {
+      "$oid": "598f59b98b237e620ade6e1a"
+    },
+    "gender": "Mens",
+    "name": "EMO 1"
+  },
+  {
+    "_id": {
+      "$oid": "598f7743bd56e76c582bd27e"
+    },
+    "club": {
+      "$oid": "598f59b9e5073c71b233abae"
+    },
+    "gender": "Mens",
+    "name": "Reading 1"
+  },
+  {
+    "_id": {
+      "$oid": "598f7743e0740cf93c040014"
+    },
+    "club": {
+      "$oid": "598f59b96081bab2d6d7409b"
+    },
+    "gender": "Mens",
+    "name": "Clapham 1"
+  },
+  {
+    "_id": {
+      "$oid": "598f7743831e5b364a04daed"
+    },
+    "club": {
+      "$oid": "598f59b99f9f02b9077fbd3a"
+    },
+    "gender": "Mens",
+    "name": "Glasgow"
+  },
+  {
+    "_id": {
+      "$oid": "598f7743b2f5f0fadc5cc217"
+    },
+    "club": {
+      "$oid": "598f59b9174e48fbae29f6ff"
+    },
+    "gender": "Mens",
+    "name": "Chevron Action Flash"
+  },
+  {
+    "_id": {
+      "$oid": "598f77436ac7d3dc44caba03"
+    },
+    "club": {
+      "$oid": "598f59b91d629a22b04a1b6e"
+    },
+    "gender": "Mens",
+    "name": "Devon 1"
+  },
+  {
+    "_id": {
+      "$oid": "598f7743c535bfbd6810c4b5"
+    },
+    "club": {
+      "$oid": "598f65d11b1ff2d2cae77028"
+    },
+    "gender": "Mens",
+    "name": "Ireland Under-24 Men"
+  },
+  {
+    "_id": {
+      "$oid": "598f77438373bcd9036c5703"
+    },
+    "club": {
+      "$oid": "598f59b97ef330aee0105791"
+    },
+    "gender": "Mens",
+    "name": "Fire of London 1"
+  },
+  {
+    "_id": {
+      "$oid": "598f77436042f07573cf38b7"
+    },
+    "club": {
+      "$oid": "598f59b9a102a38169c6715d"
+    },
+    "gender": "Mens",
+    "name": "Manchester 1"
+  },
+  {
+    "_id": {
+      "$oid": "598f7743ce82f8b8578565ed"
+    },
+    "club": {
+      "$oid": "598f65d1be2f27578e7b9294"
+    },
+    "gender": "Mens",
+    "name": "GB U24 Darwin"
+  },
+  {
+    "_id": {
+      "$oid": "598f7743b092566d1e0d759f"
+    },
+    "club": {
+      "$oid": "598f65d1be2f27578e7b9294"
+    },
+    "gender": "Mens",
+    "name": "GB U24 Brisbane"
+  },
+  {
+    "_id": {
+      "$oid": "598f7743101e384bb8e2e8ab"
+    },
+    "club": {
+      "$oid": "598f59b9921fed479776e19f"
+    },
+    "gender": "Mens",
+    "name": "SMOG"
+  },
+  {
+    "_id": {
+      "$oid": "598f7743d2aefb770be385e4"
+    },
+    "club": {
+      "$oid": "598f59b940e101c378b3b795"
+    },
+    "gender": "Mens",
+    "name": "Flump"
+  },
+  {
+    "_id": {
+      "$oid": "598f77434b27b358a8a888d0"
+    },
+    "club": {
+      "$oid": "598f59b97ea89ff25c54be80"
+    },
+    "gender": "Mens",
+    "name": "Ka-Pow!"
+  },
+  {
+    "_id": {
+      "$oid": "598f7743a2933d31d9e7bac6"
+    },
+    "club": {
+      "$oid": "598f65d1b7be43b892bff1ab"
+    },
+    "gender": "Mens",
+    "name": "Brighton City"
+  }
+]
