+++ scripts/data/tour2/teams-b-s-men.json
@@ -1,0 +1,162 @@
+[
+  {
+    "_id": {
+      "$oid": "598f7d1f7438572447c9c13e"
+    },
+    "club": {
+      "$oid": "598f65d106a45ad0f50a587f"
+    },
+    "gender": "Mens",
+    "name": "BAF"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1fa39e9867670dceb5"
+    },
+    "club": {
+      "$oid": "598f65d17474e05d2f3912e2"
+    },
+    "gender": "Mens",
+    "name": "Birmingham"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1f81fc7b6a893e6a96"
+    },
+    "club": {
+      "$oid": "598f59b97ef330aee0105791"
+    },
+    "gender": "Mens",
+    "name": "Fire of London 2"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1f4d822e9ce04479cb"
+    },
+    "club": {
+      "$oid": "598f65d1183e448af67d546f"
+    },
+    "gender": "Mens",
+    "name": "Cloud City 1"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1f715eb970d46551a3"
+    },
+    "club": {
+      "$oid": "598f65d1b7be43b892bff1ab"
+    },
+    "gender": "Mens",
+    "name": "Brighton Legends"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1f2489ead13423c4ab"
+    },
+    "club": {
+      "$oid": "598f59b91d629a22b04a1b6e"
+    },
+    "gender": "Mens",
+    "name": "Devon 2"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1f25491cd95f3451a2"
+    },
+    "club": {
+      "$oid": "598f698b552f114569c446cf"
+    },
+    "gender": "Mens",
+    "name": "Purple Cobras"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1f05bdd86eb4342eb5"
+    },
+    "club": {
+      "$oid": "598f7d3e35b61604a6fbf5a6"
+    },
+    "gender": "Mens",
+    "name": "JRX"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1f5176dcc0cb70ac16"
+    },
+    "club": {
+      "$oid": "598f65d14b27fda31d681be5"
+    },
+    "gender": "Mens",
+    "name": "Helix"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1f76fa2d509bf1c8d0"
+    },
+    "club": {
+      "$oid": "598f7d3eccb6894bac4babd7"
+    },
+    "gender": "Mens",
+    "name": "Red Leicester"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d3ece69c08753b3c9f4"
+    },
+    "club": {
+      "$oid": "598f65d1ca6ca04bf0c6d59f"
+    },
+    "gender": "Mens",
+    "name": "Camden Ultimate 1"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1f698a56ef8f9556ab"
+    },
+    "club": {
+      "$oid": "598f7d3e6742de9ba15274cc"
+    },
+    "gender": "Mens",
+    "name": "GB U20s"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1f9c2a6602eab7ea0c"
+    },
+    "club": {
+      "$oid": "598f7d3e9535467162e83fd9"
+    },
+    "gender": "Mens",
+    "name": "Irish Junior Open"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1f44c982bfa5f32976"
+    },
+    "club": {
+      "$oid": "598f65d170d0e21283133e8d"
+    },
+    "gender": "Mens",
+    "name": "Bristol 2"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1fe2e6dd22d35eb8b6"
+    },
+    "club": {
+      "$oid": "598f65d170d0e21283133e8d"
+    },
+    "gender": "Mens",
+    "name": "Bristol 1"
+  },
+  {
+    "_id": {
+      "$oid": "598f7d1fdf8c3e99a7db5bf2"
+    },
+    "club": {
+      "$oid": "598f59b98b237e620ade6e1a"
+    },
+    "gender": "Mens",
+    "name": "EMO 2"
+  }
+]
